Display UIAlertController on TVSMActionModule click action
As a first step to getting some alert-styled view displaying from the
TVSystemMenuUI, the custom Sleep Timer TVSMActionModule is displaying a
UIAlertController with a preset layout.

Next I'll try to render a custom UIView instance into the
UIAlertController view, and then place a custom layout of buttons
within that view.