Fix label initialization to match state
Previously the timer label would read 30, but the underlying state had a
value of 15. After updating the timer using the plus/minus buttons, the
label would display an unexpected value.

It seems `init` wasn't firing, and I had the UI and state initialized
with two different values.

Will probably settle on a starting value of 30, but 15 removes the need
for selecting a short timespan as I'm testing.