Sleep device when timer is lapsed
The PineBoard module is the tvOS variant of the iOS SpringBoard.

Similar to how other developers are respringing, shutting and down
rebooting iOS devices in tweaks like Activator and custom Respring
triggers, it was theorised that finding the tvOS alternative was a
good place to start.

Many different methods failed, either not sleeping the device or leaving
it in a state where the only way to turn it back on was by killing the
PineBoardServices process or hard rebooting.

Obtaining and firing a sleep action from the PineBoard System Service
proved successful and likely the proper approach, but the device would
immediately turn back on. Logs showed that permission issues were
causing the action to crash the PineBoard, which resulted in a respring
and thus an "ON" screen.

This was resolved with the inclusion of the `com.apple.background.client`
entitlement, used when code signing the bundle. An automation script has
been added to the makefile to accomplish this, which is based on
instructions referenced from the following StackOverflow answer:
https://stackoverflow.com/a/14885266/2020476